Changed percent-longs to a double. Defaulted to 400ms, 50ms, and 12.5% longs

+++ bench/src/database/results_db.go
@@ -28,7 +28,7 @@ const schemaSQL = `
 		"minWaitSeconds" INTEGER NOT NULL,
 		"longSleepMillis" INTEGER,
 		"shortSleepMillis" INTEGER,
-		"percentLongRequests" INTEGER,
+		"percentLongRequests" DOUBLE PRECISION,
 		"totalRunDurationSeconds" INTEGER NOT NULL,
 		"bestTrialID" INTEGER
     );
